Client - #117, remove nonfatal errors and warnings.
- Both Client and Server #include golem_gui_base_object.h, triggering recompiles for both.
- Ignore warning 4996 for swscanf_s deprecation, due to library header calling it.
- No functional changes.

+++ Client/Include/sp2_data_access_layer.h
@@ -284,8 +284,8 @@ namespace SP2
    struct GDataOptionsWindow
    {
       INT16 m_iColorDepth;
-      INT16 m_iResolutionX;
-      INT16 m_iResolutionY;      
+      INT32 m_iResolutionX;
+      INT32 m_iResolutionY;      
 
       REAL32 m_fMusicVolume;
       REAL32 m_fSoundVolume;
